Initial commit
Manifest is a json that setups all the code and content.js is the java script that runs with the extension. 

Right now I'm trying to figure out how to run certain parts of js on certain urls. When I run both operations on the different pages, neither of them work.

At the moment the extension works for the youtube homepage ie youtube.com
Now i need to get it to work on normal youtube videos.