Added icon and popup
Added icons and popup for the extension, and added the required manifest json code. The todo list of features for this extension is in the popup. 

Technicality wise, I'm still trying to figure out how to run the js on certain pages, and now why the button click function in background.js is not working.