Extension is working when you refresh
Extension is now working when you refresh the page. It gets rid of suggestions on youtube.com and on videos allowing users to be more productive during youtube learning.

However when a user goes from youtube.come and follows a link to a video sometimes it doesn't work, same with vice versa pathway. Need to figure out how to fix this. 

Up next is probably designing and styling the popup, then fixing this bug. Then we can add more features.